infra/analysis/requirements.txt
google-cloud-bigquery
google-cloud-datastore
pylint
yapf
